services/auth_service/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
sqlalchemy==2.0.41
passlib==1.7.4
bcrypt==3.2.2
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
email-validator==2.1.1
